refactor(cli): extract runtime evaluate handler
