+++ application/src/Nexo.CLI/Commands/Runtime/EvaluateHandler.cs
@@ -1,0 +1,186 @@
+using System.Text.Json;
+using Nexo.CLI.Runtime;
+
+namespace Nexo.CLI.Commands;
+
+internal delegate Task<RuntimeExecuteResult> RuntimeExecuteCore(
+    string goal,
+    string repoRoot,
+    string? provider,
+    bool allowMock,
+    bool runTests,
+    string testFilter,
+    string bootstrapProfile,
+    string qaPolicy,
+    string? runtimeManifestPath,
+    string? runtimeManifestJson,
+    int? maxIterationsOverride,
+    bool bootstrapApply,
+    bool bootstrapYes,
+    bool bootstrapDryRun,
+    bool runPreflight,
+    bool useHistory,
+    int historyWindow,
+    bool persistHistory,
+    string benchmarkSet,
+    bool allowVisualCapabilityDegrade,
+    CancellationToken ct);
+
+internal sealed class EvaluateHandler(RuntimeExecuteCore executeCore)
+{
+    public async Task<int> ExecuteAsync(
+        string? goalsJson,
+        string? goalsFile,
+        string policiesCsv,
+        string repoRoot,
+        string? provider,
+        bool allowMock,
+        bool runTests,
+        string testFilter,
+        string bootstrapProfile,
+        string? runtimeManifestPath,
+        string? runtimeManifestJson,
+        int? maxIterationsOverride,
+        bool bootstrapApply,
+        bool runPreflight,
+        bool useHistory,
+        int historyWindow,
+        bool persistHistory,
+        string benchmarkSet,
+        bool allowVisualCapabilityDegrade,
+        bool json,
+        CancellationToken ct)
+    {
+        var goals = ResolveGoals(goalsJson, goalsFile);
+        if (goals.Length == 0)
+        {
+            RuntimeOutputWriter.WriteEvaluateResult(new RuntimeEvaluateResult(false, "No goals provided for evaluation."), json);
+            return 1;
+        }
+
+        var policies = ResolvePolicies(policiesCsv);
+        if (policies.Length == 0)
+        {
+            RuntimeOutputWriter.WriteEvaluateResult(new RuntimeEvaluateResult(false, "No valid policies provided for evaluation."), json);
+            return 1;
+        }
+
+        var fullRepoRoot = Path.GetFullPath(string.IsNullOrWhiteSpace(repoRoot) ? Environment.CurrentDirectory : repoRoot);
+        if (!Directory.Exists(fullRepoRoot))
+        {
+            RuntimeOutputWriter.WriteEvaluateResult(new RuntimeEvaluateResult(false, $"Repo root not found: {fullRepoRoot}"), json);
+            return 1;
+        }
+
+        var scenarioResults = new List<RuntimeEvaluateScenarioResult>();
+        foreach (var goal in goals)
+        {
+            foreach (var policy in policies)
+            {
+                var run = await executeCore(
+                    goal,
+                    fullRepoRoot,
+                    provider,
+                    allowMock,
+                    runTests,
+                    testFilter,
+                    bootstrapProfile,
+                    policy,
+                    runtimeManifestPath,
+                    runtimeManifestJson,
+                    maxIterationsOverride,
+                    bootstrapApply,
+                    bootstrapYes: true,
+                    bootstrapDryRun: false,
+                    runPreflight,
+                    useHistory,
+                    historyWindow,
+                    persistHistory,
+                    benchmarkSet,
+                    allowVisualCapabilityDegrade,
+                    ct).ConfigureAwait(false);
+                scenarioResults.Add(new RuntimeEvaluateScenarioResult(
+                    AdaptiveRuntimeExecutionReport.BuildGoalPreview(goal, 80),
+                    policy,
+                    run.Ok,
+                    run.ElapsedMs,
+                    run.FailureStage,
+                    run.Summary));
+            }
+        }
+
+        var policySummary = scenarioResults
+            .GroupBy(r => r.Policy, StringComparer.OrdinalIgnoreCase)
+            .Select(g => new RuntimeEvaluatePolicySummary(
+                g.Key,
+                g.Count(),
+                g.Count(x => x.Ok),
+                g.Count(x => !x.Ok),
+                (long)Math.Round(
+                    g.Where(x => x.ElapsedMs.HasValue)
+                     .Select(x => (double)x.ElapsedMs!.Value)
+                     .DefaultIfEmpty(0d)
+                     .Average())))
+            .OrderBy(s => s.Policy, StringComparer.OrdinalIgnoreCase)
+            .ToArray();
+
+        var allPassed = scenarioResults.All(r => r.Ok);
+        var result = new RuntimeEvaluateResult(
+            allPassed,
+            allPassed
+                ? $"Evaluation complete: {scenarioResults.Count} scenario(s) passed."
+                : $"Evaluation complete: {scenarioResults.Count(r => !r.Ok)} failing scenario(s) out of {scenarioResults.Count}.",
+            scenarioResults.ToArray(),
+            policySummary);
+        RuntimeOutputWriter.WriteEvaluateResult(result, json);
+        return allPassed ? 0 : 1;
+    }
+
+
+    private static string[] ResolveGoals(string? goalsJson, string? goalsFile)
+    {
+        if (!string.IsNullOrWhiteSpace(goalsJson))
+        {
+            try
+            {
+                var parsed = JsonSerializer.Deserialize<string[]>(goalsJson);
+                return NormalizeGoals(parsed);
+            }
+            catch
+            {
+                return Array.Empty<string>();
+            }
+        }
+
+        if (!string.IsNullOrWhiteSpace(goalsFile) && File.Exists(goalsFile))
+        {
+            var lines = File.ReadAllLines(goalsFile);
+            return NormalizeGoals(lines);
+        }
+
+        return Array.Empty<string>();
+    }
+
+
+    private static string[] NormalizeGoals(IEnumerable<string>? goals)
+    {
+        return (goals ?? Array.Empty<string>())
+            .Select(g => (g ?? string.Empty).Trim())
+            .Where(g => !string.IsNullOrWhiteSpace(g))
+            .Distinct(StringComparer.OrdinalIgnoreCase)
+            .ToArray();
+    }
+
+
+    private static string[] ResolvePolicies(string csv)
+    {
+        return (csv ?? string.Empty)
+            .Split(',', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries)
+            .Select(RuntimeCommandUtilities.NormalizeQaPolicy)
+            .Where(p => p is "demo" or "release" or "prod" or "research" or "auto")
+            .Distinct(StringComparer.OrdinalIgnoreCase)
+            .ToArray();
+    }
+
+
+}
